feat(persistence): define categoria persistence entity, mapper, repository and migration

+++ src/main/java/io/github/franciscopaulinoq/zelaseveriano/domain/repository/CategoriaRepository.java
@@ -6,9 +6,9 @@ import java.util.List;
 import java.util.Optional;
 
 public interface CategoriaRepository {
-    public Optional<Categoria> findById(int categoriaId);
+    public Optional<Categoria> findById(Long categoriaId);
 
     public List<Categoria> findAll();
 
-    public boolean existsById(int categoriaId);
+    public boolean existsById(Long categoriaId);
 }
